+++ nbed/utils.py
@@ -8,14 +8,10 @@ import os
 from pathlib import Path
 from typing import Optional
 
-from openfermion import count_qubits
 from openfermion.chem.pubchem import geometry_from_pubchem
 from pydantic import ValidationError
 
-from nbed.config import NbedConfig, ProjectorEnum
-from nbed.ham_builder import HamiltonianBuilder
-
-from .driver import NbedDriver
+from nbed.config import NbedConfig
 
 logger = logging.getLogger(__name__)
 
@@ -67,7 +63,7 @@ def parse():
         "--transform",
         required=False,
         type=str,
-        help="Fermion to Qubit encoding to use e.g.'Jordan-Wigner', 'Bravyi-Kitaev' "
+        help="Fermion to Qubit encoding to use e.g.'Jordan-Wigner', 'Bravyi-Kitaev' ",
     )
     parser.add_argument(
         "--savefile",
@@ -91,143 +87,6 @@ def parse():
         logger.error(e)
 
     return config, args["transform"], args["savefile"]
-
-
-def print_summary(driver: NbedDriver, transform: str, fci: bool = False) -> None:
-    """Print a summary of the package results.
-
-    Args:
-        driver (NbedDriver): An NbedDriver to summarise.
-        transform (str): The transform used to generate a qubit Hamiltonian.
-        fci (bool): Whether to run full system fci.
-    """
-    logger.debug("Printing summary of results.")
-    # for get statements
-    default = "Not calculated."
-
-    qham = HamiltonianBuilder(
-        driver.embedded_scf,
-        constant_e_shift=driver.classical_energy,
-        transform=transform,
-    ).build()
-
-    # Would be a great place for a switch statemet when
-    # dependencies catch up with python 3.10
-    match driver.config.projector:
-        case ProjectorEnum.BOTH:
-            mu_qham, huz_qham = qham
-        case ProjectorEnum.HUZ:
-            mu_qham, huz_qham = None, qham
-        case ProjectorEnum.MU:
-            mu_qham, huz_qham = qham, None
-
-    print("".center(80, "*"))
-    logger.info("".center(80, "*"))
-    print("  Summary of Embedded Calculation".center(80))
-    logger.info("  Summary of Embedded Calculation".center(80))
-    print("".center(80, "*"))
-    logger.info("".center(80, "*"))
-
-    print(f"global (cheap) DFT calculation {driver._global_ks.e_tot}")
-    logger.info(f"global (cheap) DFT calculation {driver._global_ks.e_tot}")
-
-    if driver.config.projector in [ProjectorEnum.HUZ, ProjectorEnum.BOTH]:
-        print("".center(80, "*"))
-        logger.info("".center(80, "*"))
-        print("  Huzinaga calculation".center(20))
-        logger.info("  Huzinaga calculation".center(20))
-        print(
-            f"Total energy - active system at RHF level: {driver._huzinaga.get('e_rhf', default)}"
-        )
-        logger.info(
-            f"Total energy - active system at RHF level: {driver._huzinaga.get('e_rhf', default)}"
-        )
-        if driver.config.run_ccsd_emb is True:
-            print(
-                f"Total energy - active system at CCSD level: {driver._huzinaga.get('e_ccsd', default)}"
-            )
-            logger.info(
-                f"Total energy - active system at CCSD level: {driver._huzinaga.get('e_ccsd', default)}"
-            )
-        if driver.config.run_fci_emb is True:
-            print(
-                f"Total energy - active system at FCI level: {driver._huzinaga.get('e_fci', default)}"
-            )
-            logger.info(
-                f"Total energy - active system at FCI level: {driver._huzinaga.get('e_fci', default)}"
-            )
-
-        print(
-            f"length of huzinaga embedded fermionic Hamiltonian: {len(huz_qham.terms)}"
-        )
-        logger.info(
-            f"length of huzinaga embedded fermionic Hamiltonian: {len(huz_qham.terms)}"
-        )
-        print(f"number of qubits required: {count_qubits(huz_qham)}")
-        logger.info(f"number of qubits required: {count_qubits(huz_qham)}")
-
-    if driver.config.projector in [ProjectorEnum.MU, ProjectorEnum.BOTH]:
-        print("".center(80, "*"))
-        logger.info("".center(80, "*"))
-        print("  Mu shift calculation".center(20))
-        logger.info("  Mu shift calculation".center(20))
-        print(
-            f"Total energy - active system at RHF level: {driver._mu.get('e_rhf', default)}"
-        )
-        logger.info(
-            f"Total energy - active system at RHF level: {driver._mu.get('e_rhf', default)}"
-        )
-        if driver.config.run_ccsd_emb is True:
-            print(
-                f"Total energy - active system at CCSD level: {driver._mu.get('e_ccsd', default)}"
-            )
-            logger.info(
-                f"Total energy - active system at CCSD level: {driver._mu.get('e_ccsd', default)}"
-            )
-        if driver.config.run_fci_emb is True:
-            print(
-                f"Total energy - active system at FCI level: {driver._mu.get('e_fci', default)}"
-            )
-            logger.info(
-                f"Total energy - active system at FCI level: {driver._mu.get('e_fci', default)}"
-            )
-
-        print(f"length of mu embedded fermionic Hamiltonian: {len(mu_qham.terms)}")
-        logger.info(
-            f"length of mu embedded fermionic Hamiltonian: {len(mu_qham.terms)}"
-        )
-        print(f"number of qubits required: {count_qubits(mu_qham)}")
-        logger.info(f"number of qubits required: {count_qubits(mu_qham)}")
-
-    logger.debug("Building full system Hamiltonian for comparison.")
-    full_system_hamiltonian = HamiltonianBuilder(
-        driver._global_hf, constant_e_shift=0, transform=transform
-    ).build()
-
-    print("".center(80, "*"))
-    logger.info("".center(80, "*"))
-    print("  Summary of reference Calculation".center(80))
-    logger.info("  Summary of reference Calculation".center(80))
-    print("".center(80, "*"))
-    logger.info("".center(80, "*"))
-
-    if fci:
-        print("Running Full system FCI and preparing Hamiltonian.")
-        logger.info("Running Full system FCI and preparing Hamiltonian.")
-        print(f"Global (expensive) full FCI calculation {driver._global_fci.e_tot}")
-        logger.info(
-            f"Global (expensive) full FCI calculation {driver._global_fci.e_tot}"
-        )
-
-    print(
-        f"length of full system fermionic Hamiltonian: {len(full_system_hamiltonian.terms)}"
-    )
-
-    logger.info(
-        f"length of full system fermionic Hamiltonian: {len(full_system_hamiltonian.terms)}"
-    )
-    print(f"number of qubits required: {count_qubits(full_system_hamiltonian)}")
-    logger.info(f"number of qubits required: {count_qubits(full_system_hamiltonian)}")
 
 
 def pubchem_mol_geometry(molecule_name) -> dict: